use splitext, remove duplicate file
